sourceTag and targetTag implementation.
As discussed in email
• When a firewall is created, add a targetTag to the firewall with the
name "grid-${firewallName}". targetTag should not be exposed to the end
user via the API or console.
• When a firewall is created, allow sourceTag or sourceRange to be
specified on the firewall rule. A FirewallSourceType of FIREWALL should
be set as a sourceTag of "grid-${sourceFirewallName}" and a
FirewallSourceType of CIDR should be set as a sourceRange of the
provided value. The prefix of "grid-..." should not be exposed to the
end user.